backend/requirements.txt
fastapi
uvicorn
PyMuPDF
pytesseract
Pillow
faiss-cpu
sentence_transformers
ollama
numpy
